SERV_Preference() to take size-t in 3rd arg
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@38081 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/ncbi_servicep.h
@@ -195,9 +195,9 @@ extern NCBI_XCONNECT_EXPORT char* SERV_ServiceName(const char* service);
  * acceptable stretch for the gap (the number of candidates is n).
  */
 extern NCBI_XCONNECT_EXPORT double SERV_Preference
-(double       pref,
- double       gap,
- unsigned int n
+(double pref,
+ double gap,
+ size_t n
  );
